pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tarea-unith"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "aio-pika>=9.4.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.2.1",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "structlog>=24.1.0",
    "python-ulid>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "respx>=0.21.1",
    "testcontainers[postgresql,rabbitmq]>=4.7.0",
    "greenlet>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "receiver*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]
